+++ crates/nova_gameplay/src/sections/mod.rs
@@ -5,6 +5,7 @@ use bevy::prelude::*;
 pub mod base_section;
 pub mod controller_section;
 pub mod hull_section;
+pub mod projectile_hooks;
 pub mod thruster_section;
 pub mod torpedo_section;
 pub mod turret_section;
@@ -12,8 +13,9 @@ pub mod turret_section;
 pub mod prelude {
     pub use super::{
         base_section::prelude::*, controller_section::prelude::*, hull_section::prelude::*,
-        thruster_section::prelude::*, torpedo_section::prelude::*, turret_section::prelude::*,
-        SpaceshipRootMarker, SpaceshipSectionPlugin, SpaceshipSectionSystems,
+        projectile_hooks::prelude::*, thruster_section::prelude::*, torpedo_section::prelude::*,
+        turret_section::prelude::*, SpaceshipRootMarker, SpaceshipSectionPlugin,
+        SpaceshipSectionSystems,
     };
 }